fixes to make it all work, added DPS counter in craft menu
