+++ MagicStorage.csproj
@@ -69,6 +69,7 @@
     <Compile Include="Components\TEStoragePoint.cs" />
     <Compile Include="Components\TEStorageUnit.cs" />
     <Compile Include="CraftingGUI.cs" />
+    <Compile Include="GUIHelpers.cs" />
     <Compile Include="InterfaceHelper.cs" />
     <Compile Include="ItemData.cs" />
     <Compile Include="Items\CraftingAccess.cs" />
